Isolate platform-independant rand.c code into rand_pure.c
