+++ Makefile
@@ -36,7 +36,8 @@ ifdef SMALL
 CFLAGS += -DUSE_PRECOMPUTED_CP=0
 endif
 
-SRCS   = bignum.c ecdsa.c curves.c secp256k1.c nist256p1.c rand.c hmac.c bip32.c bip39.c pbkdf2.c base58.c base32.c
+SRCS   = bignum.c ecdsa.c curves.c secp256k1.c nist256p1.c hmac.c bip32.c bip39.c pbkdf2.c base58.c base32.c
+SRCS  += rand.c rand_pure.c
 SRCS  += address.c
 SRCS  += script.c
 SRCS  += ripemd160.c
