+++ rand.c
@@ -62,13 +62,6 @@ uint32_t random32(void)
 #endif
 }
 
-uint32_t random_uniform(uint32_t n)
-{
-	uint32_t x, max = 0xFFFFFFFF - (0xFFFFFFFF % n);
-	while ((x = random32()) >= max);
-	return x / (max / n);
-}
-
 void random_buffer(uint8_t *buf, size_t len)
 {
 #ifdef _WIN32
@@ -85,16 +78,4 @@ void random_buffer(uint8_t *buf, size_t len)
 	(void)len_read;
 	assert(len_read == len);
 #endif
-}
-
-void random_permute(char *str, size_t len)
-{
-	int i, j;
-	char t;
-	for (i = len - 1; i >= 1; i--) {
-		j = random_uniform(i + 1);
-		t = str[j];
-		str[j] = str[i];
-		str[i] = t;
-	}
 }
